Ignore appliedenergistics2-core in the modlist
Applied Energistics 2's core mod (modID "appliedenergistics2-core") prints
a different string on each launch of Minecraft to
`Loader.instance().getIndexedModList()`
This patch will ignore this specific core mod when comparing client and
server modlists. Since this mod is included in the jar of AE2, this
coremod will update correctly when AE2 is updated on the server.