Temporary fix for the windows issue with .deleteonexit()
Saves a reference file in config/serversync that is used to delete the
specified mods/configs when minecraft is loaded next.
Rather dirty as I could not find a way to get forge to restart its
Loader so the game needs to be terminated before FML complains about
missing the deleted mods, which in turn requires the user to restart
minecraft again without any real prompt to do so.